+++ Razor.Blade/Html5/GeneratedHead.cs
@@ -7,7 +7,7 @@ using ToSic.Razor.Internals.Documentation;
 // ****
 // This is auto-generated code - don't modify
 // Re-run the generation program to recreate
-// Created 06.12.2022 17:29
+// Created 06.12.2022 18:50
 //
 // Each tag and attributes of it prepare code, and they return an object of the same type again
 // to allow fluid chaining of the commands
@@ -37,7 +37,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;base&gt; tag with optional contents
     /// </summary>
     
-    internal Base() : base("base", new TagOptions { Close = false })
+    internal Base() : base("base", new TagOptions(close: false))
     {
     }
     
@@ -49,7 +49,7 @@ namespace ToSic.Razor.Html5
     /// A Base object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Base(params object[] content) : base("base", new TagOptions { Close = false }, content)
+    internal Base(params object[] content) : base("base", new TagOptions(close: false), content)
     {
     }
     
@@ -122,7 +122,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;meta&gt; tag with optional contents
     /// </summary>
     
-    internal Meta() : base("meta", new TagOptions { Close = false })
+    internal Meta() : base("meta", new TagOptions(close: false))
     {
     }
     
@@ -134,7 +134,7 @@ namespace ToSic.Razor.Html5
     /// A Meta object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Meta(params object[] content) : base("meta", new TagOptions { Close = false }, content)
+    internal Meta(params object[] content) : base("meta", new TagOptions(close: false), content)
     {
     }