+++ Razor.Blade/Html5/GeneratedMedia.cs
@@ -7,7 +7,7 @@ using ToSic.Razor.Internals.Documentation;
 // ****
 // This is auto-generated code - don't modify
 // Re-run the generation program to recreate
-// Created 06.12.2022 17:29
+// Created 06.12.2022 18:50
 //
 // Each tag and attributes of it prepare code, and they return an object of the same type again
 // to allow fluid chaining of the commands
@@ -163,7 +163,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;source&gt; tag with optional contents
     /// </summary>
     
-    internal Source() : base("source", new TagOptions { Close = false })
+    internal Source() : base("source", new TagOptions(close: false))
     {
     }
     
@@ -175,7 +175,7 @@ namespace ToSic.Razor.Html5
     /// A Source object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Source(params object[] content) : base("source", new TagOptions { Close = false }, content)
+    internal Source(params object[] content) : base("source", new TagOptions(close: false), content)
     {
     }
     
@@ -261,7 +261,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;track&gt; tag with optional contents
     /// </summary>
     
-    internal Track() : base("track", new TagOptions { Close = false })
+    internal Track() : base("track", new TagOptions(close: false))
     {
     }
     
@@ -273,7 +273,7 @@ namespace ToSic.Razor.Html5
     /// A Track object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Track(params object[] content) : base("track", new TagOptions { Close = false }, content)
+    internal Track(params object[] content) : base("track", new TagOptions(close: false), content)
     {
     }