+++ Razor.Blade/Html5/GeneratedSimple.cs
@@ -7,7 +7,7 @@ using ToSic.Razor.Internals.Documentation;
 // ****
 // This is auto-generated code - don't modify
 // Re-run the generation program to recreate
-// Created 06.12.2022 17:29
+// Created 06.12.2022 18:50
 //
 // Each tag and attributes of it prepare code, and they return an object of the same type again
 // to allow fluid chaining of the commands
@@ -37,7 +37,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;br&gt; tag with optional contents
     /// </summary>
     
-    internal Br() : base("br", new TagOptions { Close = false })
+    internal Br() : base("br", new TagOptions(close: false))
     {
     }
     
@@ -49,7 +49,7 @@ namespace ToSic.Razor.Html5
     /// A Br object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Br(params object[] content) : base("br", new TagOptions { Close = false }, content)
+    internal Br(params object[] content) : base("br", new TagOptions(close: false), content)
     {
     }
     
@@ -347,7 +347,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;hr&gt; tag with optional contents
     /// </summary>
     
-    internal Hr() : base("hr", new TagOptions { Close = false })
+    internal Hr() : base("hr", new TagOptions(close: false))
     {
     }
     
@@ -359,7 +359,7 @@ namespace ToSic.Razor.Html5
     /// A Hr object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Hr(params object[] content) : base("hr", new TagOptions { Close = false }, content)
+    internal Hr(params object[] content) : base("hr", new TagOptions(close: false), content)
     {
     }
     
@@ -502,7 +502,7 @@ namespace ToSic.Razor.Html5
     /// Generate an &lt;wbr&gt; tag with optional contents
     /// </summary>
     
-    internal Wbr() : base("wbr", new TagOptions { Close = false })
+    internal Wbr() : base("wbr", new TagOptions(close: false))
     {
     }
     
@@ -514,7 +514,7 @@ namespace ToSic.Razor.Html5
     /// A Wbr object with all the attributes available in that tag
     /// </returns>
     /// <param name="content">one or more objects (strings or tags) which will be inside the tag</param>
-    internal Wbr(params object[] content) : base("wbr", new TagOptions { Close = false }, content)
+    internal Wbr(params object[] content) : base("wbr", new TagOptions(close: false), content)
     {
     }